refactor: reorganize config structure and update import paths for constants
