+++ app/routes/_authenticated+/tasks+/_index/components/columns.tsx
@@ -4,7 +4,7 @@ import { Badge } from '~/components/ui/badge'
 import { Checkbox } from '~/components/ui/checkbox'
 import { labels } from '../../_shared/data/data'
 import type { Task } from '../../_shared/data/schema'
-import { FILTER_FIELD_LABELS } from '../config'
+import { FILTER_FIELD_LABELS } from '../config/constants'
 import { DataTableColumnHeader } from './data-table-column-header'
 import { DataTableRowActions } from './data-table-row-actions'
