+++ app/routes/_authenticated+/tasks+/_index/components/data-table-faceted-filter.tsx
@@ -18,7 +18,8 @@ import {
 } from '~/components/ui/popover'
 import { Separator } from '~/components/ui/separator'
 import { cn } from '~/lib/utils'
-import { useDataTableState, type Filters } from '../hooks/use-data-table-state'
+import type { Filters } from '../config'
+import { useDataTableState } from '../hooks/use-data-table-state'
 
 interface DataTableFacetedFilterProps {
   title?: string
