+++ app/routes/_authenticated+/tasks+/_index/components/data-table-pagination.tsx
@@ -13,7 +13,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from '~/components/ui/select'
-import { PAGINATION_PER_PAGE_ITEMS } from '../config'
+import { PAGINATION_PER_PAGE_ITEMS } from '../config/constants'
 import { useDataTableState } from '../hooks/use-data-table-state'
 
 interface DataTablePaginationProps<TData> {
