+++ app/routes/_authenticated+/tasks+/_index/config/constants.ts
@@ -9,6 +9,7 @@ import {
   IconStopwatch,
 } from '@tabler/icons-react'
 
+// Faceted filter options
 export const FILTER_FIELD_LABELS = {
   status: [
     {
@@ -55,9 +56,6 @@ export const FILTER_FIELD_LABELS = {
     },
   ],
 } as const
-export const PAGINATION_PER_PAGE_ITEMS = ['10', '20', '30', '40', '50'] as const
 
-export const FILTER_FIELDS = Object.keys(
-  FILTER_FIELD_LABELS,
-) as (keyof typeof FILTER_FIELD_LABELS)[]
-export const PAGINATION_PER_PAGE_DEFAULT = PAGINATION_PER_PAGE_ITEMS[0]
+// Pagination items per page options
+export const PAGINATION_PER_PAGE_ITEMS = ['10', '20', '30', '40', '50'] as const
