+++ app/routes/_authenticated+/tasks+/_index/config/index.ts
@@ -1,0 +1,3 @@
+export * from './constants'
+export * from './schema'
+export * from './types'
